Moved YAMLModeWrapper template class (and all of its subclasses) out of modellistfromyamlfile.cpp, and into their own header and module (yamlmodelwrapper.h and yamlmodelwrapper.cpp). Didn't change anything, just moved stuff!
